commit to reverseInPlace and reversed
The reverseInPlace was changing the original array first and exacting the element from the array, so it will be duplicated. I create a new array as the medium and make a deep copy of it to the old array. For reversed method, it returns the unchanged array instead of the new array.